# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JogjaSplorasi API",
    "version": "7.0.0",
    "description": "Phase 7 API documentation for JogjaSplorasi backend."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness health check",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Database readiness check",
        "responses": {
          "200": {
            "description": "Ready"
          },
          "503": {
            "description": "Not ready"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "email",
                  "password"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "responses": {
          "200": {
            "description": "Logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/me/profile": {
      "patch": {
        "summary": "Update profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
    "/api/me/preferences": {
      "get": {
        "summary": "Get preferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "summary": "Update preferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
    "/api/destinations": {
      "get": {
        "summary": "List destinations",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "TOURISM",
                "CULINARY",
                "CULTURE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/recommendations/today": {
      "get": {
        "summary": "Get 3 daily recommendation cards",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quiz/start": {
      "post": {
        "summary": "Start quiz attempt",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Started"
          }
        }
      }
    },
    "/api/quiz/submit": {
      "post": {
        "summary": "Submit quiz attempt",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Submitted"
          }
        }
      }
    },
    "/api/stats/me/summary": {
      "get": {
        "summary": "User profile summary stats",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/destinations": {
      "post": {
        "summary": "Create destination as admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/ai/guide/chat": {
      "post": {
        "summary": "Chat with database-grounded AI tour guide",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI guide answer with cited destinations"
          }
        }
      }
    },
    "/api/ai/itinerary/suggest": {
      "post": {
        "summary": "Generate itinerary suggestions from verified destinations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Suggested itinerary plan"
          }
        }
      }
    }
  }
}